Crate google_logging1_beta3

This documentation was generated from logging crate version 0.1.8+20150326, where 20150326 is the exact revision of the logging:v1beta3 schema built by the mako code generator v0.1.8. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

projects

log services indexes list, log services list, log services sinks create, log services sinks delete, log services sinks get, log services sinks list, log services sinks update, logs delete, logs entries write, logs list, logs sinks create, logs sinks delete, logs sinks get, logs sinks list and logs sinks update

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.projects().logs_sinks_get(...).doit()
let r = hub.projects().log_services_sinks_update(...).doit()
let r = hub.projects().logs_sinks_update(...).doit()
let r = hub.projects().log_services_sinks_create(...).doit()
let r = hub.projects().logs_sinks_create(...).doit()
let r = hub.projects().log_services_sinks_get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-logging1_beta3 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_logging1_beta3 as logging1_beta3;
use logging1_beta3::LogSink;
use logging1_beta3::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use logging1_beta3::Logging;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Logging::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = LogSink::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.projects().log_services_sinks_update(req, "projectsId", "logServicesId", "sinksId")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); }

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

ListLogServiceIndexesResponse
Result returned from ListLogServiceIndexesRequest.

ListLogServiceSinksResponse
Result returned from ListLogServiceSinks.

ListLogServicesResponse
Result returned from ListLogServicesRequest.

ListLogSinksResponse
Result returned from ListLogSinks.

ListLogsResponse
Result returned from ListLogs.

Log
A log object.

LogEntry
An individual entry in a log.

LogEntryMetadata
Additional data that is associated with a log entry, set by the service creating the log entry.

LogError
A problem in a sink or the sink's configuration.

LogService
A log service object.

LogSink
An object that describes where a log may be written.

Logging
Central instance to access all Logging related resource activities

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

ProjectLogDeleteCall
Deletes the specified log resource and all log entries contained in it.

ProjectLogEntryWriteCall
Creates one or more log entries in a log. You must supply a list of LogEntry objects, named entries. Each LogEntry object must contain a payload object and a LogEntryMetadata object that describes the entry. You must fill in all the fields of the entry, metadata, and payload. You can also supply a map, commonLabels, that supplies default (key, value) data for the entries[].metadata.labels maps, saving you the trouble of creating identical copies for each entry.

ProjectLogListCall
Lists log resources belonging to the specified project.

ProjectLogServiceIndexeListCall
Lists log service indexes associated with a log service.

ProjectLogServiceListCall
Lists log services associated with log entries ingested for a project.

ProjectLogServiceSinkCreateCall
Creates the specified log service sink resource.

ProjectLogServiceSinkDeleteCall
Deletes the specified log service sink.

ProjectLogServiceSinkGetCall
Gets the specified log service sink resource.

ProjectLogServiceSinkListCall
Lists log service sinks associated with the specified service.

ProjectLogServiceSinkUpdateCall
Creates or update the specified log service sink resource.

ProjectLogSinkCreateCall
Creates the specified log sink resource.

ProjectLogSinkDeleteCall
Deletes the specified log sink resource.

ProjectLogSinkGetCall
Gets the specified log sink resource.

ProjectLogSinkListCall
Lists log sinks associated with the specified log.

ProjectLogSinkUpdateCall
Creates or updates the specified log sink resource.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Logging hub.

Status
Represents the RPC error status for Google APIs. See http://go/errormodel for details.

WriteLogEntriesRequest
The parameters to WriteLogEntries.

WriteLogEntriesResponse
Result returned from WriteLogEntries. empty

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.