Struct google_logging1_beta3::HttpRequest

pub struct HttpRequest {
    pub status: Option<i32>,
    pub request_method: Option<String>,
    pub request_url: Option<String>,
    pub cache_hit: Option<bool>,
    pub referer: Option<String>,
    pub remote_ip: Option<String>,
    pub user_agent: Option<String>,
    pub validated_with_origin_server: Option<bool>,
    pub request_size: Option<String>,
    pub response_size: Option<String>,
}

A common proto for logging HTTP requests.

This type is not used in any activity, and only used as part of another schema.

Fields

status

A response code indicates the status of response, e.g., 200.

request_method

Request method, such as GET, HEAD, PUT or POST.

request_url

Contains the scheme (http|https), the host name, the path and the query portion of the URL that was requested.

cache_hit

Whether or not an entity was served from cache (with or without validation).

referer

Referer (a.k.a. referrer) URL of request, as defined in http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html.

remote_ip

IP address of the client who issues the HTTP request. Could be either IPv4 or IPv6.

user_agent

User agent sent by the client, e.g., "Mozilla/4.0 (compatible; MSIE 6.0; Windows 98; Q312461; .NET CLR 1.0.3705)".

validated_with_origin_server

Whether or not the response was validated with the origin server before being served from cache. This field is only meaningful if cache_hit is True.

request_size

Size of the HTTP request message in bytes, including request headers and the request body.

response_size

Size of the HTTP response message in bytes sent back to the client, including response headers and response body.