Struct google_logging1_beta3::LogSink

pub struct LogSink {
    pub destination: Option<String>,
    pub errors: Option<Vec<LogError>>,
    pub name: Option<String>,
}

An object that describes where a log may be written.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

logs sinks get projects (response)

log services sinks update projects (request|response)

logs sinks update projects (request|response)

log services sinks create projects (request|response)

logs sinks create projects (request|response)

log services sinks get projects (response)

Fields

destination

The resource to send log entries to. The supported sink resource types are: + Google Cloud Storage: storage.googleapis.com/BUCKET or BUCKET.storage.googleapis.com/ + Google BigQuery: bigquery.googleapis.com/projects/PROJECT/datasets/DATASET Currently the Cloud Logging API supports at most one sink for each resource type per log or log service resource.

errors

Output only. All active errors found for this sink.

name

The name of this sink. This is a client-assigned identifier for the resource. This is ignored by UpdateLogSink and UpdateLogServicesSink.