Struct google_logging1_beta3::ProjectLogSinkDeleteCall

pub struct ProjectLogSinkDeleteCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Deletes a log sink. After deletion, no new log entries are written to the destination.

A builder for the logs.sinks.delete method supported by a project resource. It is not used directly, but through a ProjectMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.projects().logs_sinks_delete("projectsId", "logsId", "sinksId")
             .doit();

Methods

impl<'a, C, A> ProjectLogSinkDeleteCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, Empty)>

Perform the operation you have build so far.

fn projects_id(self, new_value: &str) -> ProjectLogSinkDeleteCall<'a, C, A>

Part of sinkName. The resource name of the log sink to delete.

Sets the projects id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn logs_id(self, new_value: &str) -> ProjectLogSinkDeleteCall<'a, C, A>

Part of sinkName. See documentation of projectsId.

Sets the logs id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn sinks_id(self, new_value: &str) -> ProjectLogSinkDeleteCall<'a, C, A>

Part of sinkName. See documentation of projectsId.

Sets the sinks id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn delegate(self, new_value: &'a mut Delegate) -> ProjectLogSinkDeleteCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> ProjectLogSinkDeleteCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

bearer_token (query-string) - OAuth bearer token.

pp (query-boolean) - Pretty-print response.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

upload_protocol (query-string) - Upload protocol for media (e.g. "raw", "multipart").

access_token (query-string) - OAuth access token.

fields (query-string) - Selector specifying which fields to include in a partial response.

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.

callback (query-string) - JSONP

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

uploadType (query-string) - Legacy upload protocol for media (e.g. "media", "multipart").

alt (query-string) - Data format for response.

$.xgafv (query-string) - V1 error format.

fn add_scope<T>(self, scope: T) -> ProjectLogSinkDeleteCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::CloudPlatform.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.