Struct google_logging1_beta3::ProjectMethods

pub struct ProjectMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on project resources. It is not used directly, but through the Logging hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_logging1_beta3 as logging1_beta3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use logging1_beta3::Logging;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Logging::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `log_services_indexes_list(...)`, `log_services_list(...)`, `log_services_sinks_create(...)`, `log_services_sinks_delete(...)`, `log_services_sinks_get(...)`, `log_services_sinks_list(...)`, `log_services_sinks_update(...)`, `logs_delete(...)`, `logs_entries_write(...)`, `logs_list(...)`, `logs_sinks_create(...)`, `logs_sinks_delete(...)`, `logs_sinks_get(...)`, `logs_sinks_list(...)` and `logs_sinks_update(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

fn log_services_list(&self, projects_id: &str) -> ProjectLogServiceListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists log services associated with log entries ingested for a project.

Arguments

projectsId - Part of projectName. The project resource whose services are to be listed.

fn logs_list(&self, projects_id: &str) -> ProjectLogListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists log resources belonging to the specified project.

Arguments

projectsId - Part of projectName. The project name for which to list the log resources.

fn logs_sinks_get(&self, projects_id: &str, logs_id: &str, sinks_id: &str) -> ProjectLogSinkGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified log sink resource.

Arguments

projectsId - Part of sinkName. The name of the sink resource to return.

logsId - Part of sinkName. See documentation of projectsId.

sinksId - Part of sinkName. See documentation of projectsId.

fn log_services_sinks_update(&self, request: LogSink, projects_id: &str, log_services_id: &str, sinks_id: &str) -> ProjectLogServiceSinkUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates or update the specified log service sink resource.

Arguments

request - No description provided.

projectsId - Part of sinkName. The name of the sink to update.

logServicesId - Part of sinkName. See documentation of projectsId.

sinksId - Part of sinkName. See documentation of projectsId.

fn logs_entries_write(&self, request: WriteLogEntriesRequest, projects_id: &str, logs_id: &str) -> ProjectLogEntryWriteCall<'a, C, A>

Create a builder to help you perform the following task:

Creates one or more log entries in a log. You must supply a list of LogEntry objects, named entries. Each LogEntry object must contain a payload object and a LogEntryMetadata object that describes the entry. You must fill in all the fields of the entry, metadata, and payload. You can also supply a map, commonLabels, that supplies default (key, value) data for the entries[].metadata.labels maps, saving you the trouble of creating identical copies for each entry.

Arguments

request - No description provided.

projectsId - Part of logName. The name of the log resource into which to insert the log entries.

logsId - Part of logName. See documentation of projectsId.

fn log_services_sinks_delete(&self, projects_id: &str, log_services_id: &str, sinks_id: &str) -> ProjectLogServiceSinkDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified log service sink.

Arguments

projectsId - Part of sinkName. The name of the sink to delete.

logServicesId - Part of sinkName. See documentation of projectsId.

sinksId - Part of sinkName. See documentation of projectsId.

fn log_services_indexes_list(&self, projects_id: &str, log_services_id: &str) -> ProjectLogServiceIndexeListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists log service indexes associated with a log service.

Arguments

projectsId - Part of serviceName. A log service resource of the form /projects/*/logServices/*. The service indexes of the log service are returned. Example: "/projects/myProj/logServices/appengine.googleapis.com".

logServicesId - Part of serviceName. See documentation of projectsId.

fn logs_sinks_list(&self, projects_id: &str, logs_id: &str) -> ProjectLogSinkListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists log sinks associated with the specified log.

Arguments

projectsId - Part of logName. The log for which to list sinks.

logsId - Part of logName. See documentation of projectsId.

fn logs_sinks_update(&self, request: LogSink, projects_id: &str, logs_id: &str, sinks_id: &str) -> ProjectLogSinkUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates or updates the specified log sink resource.

Arguments

request - No description provided.

projectsId - Part of sinkName. The name of the sink to update.

logsId - Part of sinkName. See documentation of projectsId.

sinksId - Part of sinkName. See documentation of projectsId.

fn log_services_sinks_create(&self, request: LogSink, projects_id: &str, log_services_id: &str) -> ProjectLogServiceSinkCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates the specified log service sink resource.

Arguments

request - No description provided.

projectsId - Part of serviceName. The name of the service in which to create a sink.

logServicesId - Part of serviceName. See documentation of projectsId.

fn logs_sinks_delete(&self, projects_id: &str, logs_id: &str, sinks_id: &str) -> ProjectLogSinkDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified log sink resource.

Arguments

projectsId - Part of sinkName. The name of the sink to delete.

logsId - Part of sinkName. See documentation of projectsId.

sinksId - Part of sinkName. See documentation of projectsId.

fn logs_sinks_create(&self, request: LogSink, projects_id: &str, logs_id: &str) -> ProjectLogSinkCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates the specified log sink resource.

Arguments

request - No description provided.

projectsId - Part of logName. The log in which to create a sink resource.

logsId - Part of logName. See documentation of projectsId.

fn log_services_sinks_get(&self, projects_id: &str, log_services_id: &str, sinks_id: &str) -> ProjectLogServiceSinkGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified log service sink resource.

Arguments

projectsId - Part of sinkName. The name of the sink to return.

logServicesId - Part of sinkName. See documentation of projectsId.

sinksId - Part of sinkName. See documentation of projectsId.

fn log_services_sinks_list(&self, projects_id: &str, log_services_id: &str) -> ProjectLogServiceSinkListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists log service sinks associated with the specified service.

Arguments

projectsId - Part of serviceName. The name of the service for which to list sinks.

logServicesId - Part of serviceName. See documentation of projectsId.

fn logs_delete(&self, projects_id: &str, logs_id: &str) -> ProjectLogDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified log resource and all log entries contained in it.

Arguments

projectsId - Part of logName. The log resource to delete.

logsId - Part of logName. See documentation of projectsId.