Struct google_logging1_beta3::LogEntry

pub struct LogEntry {
    pub proto_payload: Option<HashMap<String, String>>,
    pub log: Option<String>,
    pub text_payload: Option<String>,
    pub insert_id: Option<String>,
    pub struct_payload: Option<HashMap<String, String>>,
    pub metadata: Option<LogEntryMetadata>,
}

An individual entry in a log.

This type is not used in any activity, and only used as part of another schema.

Fields

proto_payload

The log entry payload, represented as a protocol buffer that is expressed as a JSON object. You can only pass protoPayload values that belong to a set of approved types.

log

The log to which this entry belongs. When a log entry is ingested, the value of this field is set by the logging system.

text_payload

The log entry payload, represented as a text string.

insert_id

A unique ID for the log entry. If you provide this field, the logging service considers other log entries in the same log with the same ID as duplicates which can be removed.

struct_payload

The log entry payload, represented as a structure that is expressed as a JSON object.

metadata

Information about the log entry.