Struct google_logging1_beta3::ProjectMethods

pub struct ProjectMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on project resources. It is not used directly, but through the Logging hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_logging1_beta3 as logging1_beta3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use logging1_beta3::Logging;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Logging::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `log_services_indexes_list(...)`, `log_services_list(...)`, `log_services_sinks_create(...)`, `log_services_sinks_delete(...)`, `log_services_sinks_get(...)`, `log_services_sinks_list(...)`, `log_services_sinks_update(...)`, `logs_delete(...)`, `logs_entries_write(...)`, `logs_list(...)`, `logs_sinks_create(...)`, `logs_sinks_delete(...)`, `logs_sinks_get(...)`, `logs_sinks_list(...)`, `logs_sinks_update(...)`, `metrics_create(...)`, `metrics_delete(...)`, `metrics_get(...)`, `metrics_list(...)`, `metrics_update(...)`, `sinks_create(...)`, `sinks_delete(...)`, `sinks_get(...)`, `sinks_list(...)` and `sinks_update(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

fn logs_delete(&self, projects_id: &str, logs_id: &str) -> ProjectLogDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a log and all its log entries. The log will reappear if it receives new entries.

Arguments

projectsId - Part of logName. The resource name of the log to be deleted.

logsId - Part of logName. See documentation of projectsId.

fn logs_sinks_create(&self, request: LogSink, projects_id: &str, logs_id: &str) -> ProjectLogSinkCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a log sink. All log entries for a specified log are written to the destination.

Arguments

request - No description provided.

projectsId - Part of logName. The resource name of the log to which to the sink is bound.

logsId - Part of logName. See documentation of projectsId.

fn metrics_get(&self, projects_id: &str, metrics_id: &str) -> ProjectMetricGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get the specified log metric resource.

Arguments

projectsId - Part of metricName. The resource name of the desired metric.

metricsId - Part of metricName. See documentation of projectsId.

fn logs_list(&self, projects_id: &str) -> ProjectLogListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the logs in the project. Only logs that have entries are listed.

Arguments

projectsId - Part of projectName. The resource name of the project whose logs are requested. If both serviceName and serviceIndexPrefix are empty, then all logs with entries in this project are listed.

fn metrics_create(&self, request: LogMetric, projects_id: &str) -> ProjectMetricCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Create the specified log metric resource.

Arguments

request - No description provided.

projectsId - Part of projectName. The resource name of the project in which to create the metric.

fn sinks_get(&self, projects_id: &str, sinks_id: &str) -> ProjectSinkGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a project sink.

Arguments

projectsId - Part of sinkName. The resource name of the project sink to return.

sinksId - Part of sinkName. See documentation of projectsId.

fn logs_sinks_get(&self, projects_id: &str, logs_id: &str, sinks_id: &str) -> ProjectLogSinkGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a log sink.

Arguments

projectsId - Part of sinkName. The resource name of the log sink to return.

logsId - Part of sinkName. See documentation of projectsId.

sinksId - Part of sinkName. See documentation of projectsId.

fn log_services_sinks_update(&self, request: LogSink, projects_id: &str, log_services_id: &str, sinks_id: &str) -> ProjectLogServiceSinkUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a log service sink. If the sink does not exist, it is created.

Arguments

request - No description provided.

projectsId - Part of sinkName. The resource name of the log service sink to update.

logServicesId - Part of sinkName. See documentation of projectsId.

sinksId - Part of sinkName. See documentation of projectsId.

fn sinks_update(&self, request: LogSink, projects_id: &str, sinks_id: &str) -> ProjectSinkUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a project sink. If the sink does not exist, it is created. The destination, filter, or both may be updated.

Arguments

request - No description provided.

projectsId - Part of sinkName. The resource name of the project sink to update.

sinksId - Part of sinkName. See documentation of projectsId.

fn log_services_indexes_list(&self, projects_id: &str, log_services_id: &str) -> ProjectLogServiceIndexeListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the current index values for a log service.

Arguments

projectsId - Part of serviceName. The resource name of a log service whose service indexes are requested. Example: "projects/my-project-id/logServices/appengine.googleapis.com".

logServicesId - Part of serviceName. See documentation of projectsId.

fn sinks_list(&self, projects_id: &str) -> ProjectSinkListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists project sinks associated with a project.

Arguments

projectsId - Part of projectName. The project whose sinks are wanted.

fn log_services_list(&self, projects_id: &str) -> ProjectLogServiceListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the log services that have log entries in this project.

Arguments

projectsId - Part of projectName. The resource name of the project whose services are to be listed.

fn logs_entries_write(&self, request: WriteLogEntriesRequest, projects_id: &str, logs_id: &str) -> ProjectLogEntryWriteCall<'a, C, A>

Create a builder to help you perform the following task:

Writes log entries to Cloud Logging. Each entry consists of a LogEntry object. You must fill in all the fields of the object, including one of the payload fields. You may supply a map, commonLabels, that holds default (key, value) data for the entries[].metadata.labels map in each entry, saving you the trouble of creating identical copies for each entry.

Arguments

request - No description provided.

projectsId - Part of logName. The resource name of the log that will receive the log entries.

logsId - Part of logName. See documentation of projectsId.

fn log_services_sinks_delete(&self, projects_id: &str, log_services_id: &str, sinks_id: &str) -> ProjectLogServiceSinkDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a log service sink. After deletion, no new log entries are written to the destination.

Arguments

projectsId - Part of sinkName. The resource name of the log service sink to delete.

logServicesId - Part of sinkName. See documentation of projectsId.

sinksId - Part of sinkName. See documentation of projectsId.

fn sinks_delete(&self, projects_id: &str, sinks_id: &str) -> ProjectSinkDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a project sink. After deletion, no new log entries are written to the destination.

Arguments

projectsId - Part of sinkName. The resource name of the project sink to delete.

sinksId - Part of sinkName. See documentation of projectsId.

fn log_services_sinks_create(&self, request: LogSink, projects_id: &str, log_services_id: &str) -> ProjectLogServiceSinkCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a log service sink. All log entries from a specified log service are written to the destination.

Arguments

request - No description provided.

projectsId - Part of serviceName. The resource name of the log service to which the sink is bound.

logServicesId - Part of serviceName. See documentation of projectsId.

fn logs_sinks_delete(&self, projects_id: &str, logs_id: &str, sinks_id: &str) -> ProjectLogSinkDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a log sink. After deletion, no new log entries are written to the destination.

Arguments

projectsId - Part of sinkName. The resource name of the log sink to delete.

logsId - Part of sinkName. See documentation of projectsId.

sinksId - Part of sinkName. See documentation of projectsId.

fn sinks_create(&self, request: LogSink, projects_id: &str) -> ProjectSinkCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a project sink. A logs filter determines which log entries are written to the destination.

Arguments

request - No description provided.

projectsId - Part of projectName. The resource name of the project to which the sink is bound.

fn log_services_sinks_get(&self, projects_id: &str, log_services_id: &str, sinks_id: &str) -> ProjectLogServiceSinkGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a log service sink.

Arguments

projectsId - Part of sinkName. The resource name of the log service sink to return.

logServicesId - Part of sinkName. See documentation of projectsId.

sinksId - Part of sinkName. See documentation of projectsId.

fn logs_sinks_list(&self, projects_id: &str, logs_id: &str) -> ProjectLogSinkListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists log sinks associated with a log.

Arguments

projectsId - Part of logName. The log whose sinks are wanted. For example, "compute.google.com/syslog".

logsId - Part of logName. See documentation of projectsId.

fn metrics_update(&self, request: LogMetric, projects_id: &str, metrics_id: &str) -> ProjectMetricUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Create or update the specified log metric resource.

Arguments

request - No description provided.

projectsId - Part of metricName. The resource name of the metric to update.

metricsId - Part of metricName. See documentation of projectsId.

fn logs_sinks_update(&self, request: LogSink, projects_id: &str, logs_id: &str, sinks_id: &str) -> ProjectLogSinkUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a log sink. If the sink does not exist, it is created.

Arguments

request - No description provided.

projectsId - Part of sinkName. The resource name of the sink to update.

logsId - Part of sinkName. See documentation of projectsId.

sinksId - Part of sinkName. See documentation of projectsId.

fn metrics_delete(&self, projects_id: &str, metrics_id: &str) -> ProjectMetricDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified log metric.

Arguments

projectsId - Part of metricName. The resource name of the metric to delete.

metricsId - Part of metricName. See documentation of projectsId.

fn log_services_sinks_list(&self, projects_id: &str, log_services_id: &str) -> ProjectLogServiceSinkListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists log service sinks associated with a log service.

Arguments

projectsId - Part of serviceName. The log service whose sinks are wanted.

logServicesId - Part of serviceName. See documentation of projectsId.

fn metrics_list(&self, projects_id: &str) -> ProjectMetricListCall<'a, C, A>

Create a builder to help you perform the following task:

List log metrics associated with the specified project.

Arguments

projectsId - Part of projectName. The resource name for the project whose metrics are wanted.